% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Title fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Number of a state}

\item{year}{The year in question}
}
\value{
A plot or set of plots based on latitude and longitude from the data file
}
\description{
Accepts a state number and year from user and write the filename using the year and the make_filename function
gets a data frame from fars_read(), Error checks to make sure the state number exists
If so, uses maps and graphics to create plots based on latitude and longitude from the data file
}
\examples{
\dontrun{
  fars_map_state(2, 2014)
}


}
