% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Title fars_read}
\source{
extdate/accident_xxxx.csv.bz2
}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a character of filename of a csv file to read}
}
\value{
a tbl_df with the data of the CSV file.
    The file must exist, because an error message will be show if isnt
}
\description{
This function allows a CSV extension file and returns a tbl_df with the data
If the parameter filnename does not exist, return an error
}
\examples{
\dontrun{x <- fars_read('myFile.csv')}
\dontrun{system.file("extdata", "accident_xxxx.csv.bz2", package = "packFars") }

}
