% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{#' Title fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{one or a list of years}
}
\value{
Creates one or more datasets based on year number.  Returns NULL if there is an error
}
\description{
This function accepts one or a list of years, calls the previous function make_filename()
and with each of the years of the list will run, and  mutate a new column with the year and then select
the columns MONTH and year, if the file does not exist an error will be thrown.
}
\details{
Uses make_filename(year)
     fars_read(file)
}
\examples{
\dontrun{
     fars_read_years(as.list(2013, 2014, 2015))
}

}
