% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Title fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{One or more years, no error checking}
}
\value{
A wide data frame of counts by month and year,
}
\description{
Accepts one or a list of years and use the function fars_read_years().
Uses various dplyr functions count the number of observations by month for a year.
}
\examples{
\dontrun{
  fars_summarize_years(c(2013, 2014))
}

}
