% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FarsPackages.R
\docType{package}
\name{FarsPackages}
\alias{FarsPackages}
\title{FARS Packages}
\description{
This group of functions is to document and visualize the fatal accident registry,
which is in the national census that provides the United States public with annual
data on fatal injuries sustained in automobile accidents.
}
\author{
Nelson Jair Centeno Moreno
}
