% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Title make_filename}
\source{
extdate/accident_xxxx.csv.bz2
}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year in the name of the file, no format Restriction}
}
\value{
a filename in the format 'accident_xxxx.csv.bz2'.
}
\description{
This function allows the user to create a string accident data file name by year.
bz2 files are compressed files.
}
\note{
This package does not depends on any extensions
}
\examples{
newFileName <- make_filename('2017')
\dontrun{system.file("extdata", "accident_xxxx.csv.bz2", package = "packFars")}

}
